test: Show number of tests left to be executed (#6588)
When running tests locally on many browsers, it can be difficult to tell
which browsers are done. You need to look at the number of executed and
skipped tests, sum them, and compare that to the total number of tests.
With each browser skipping different numbers of tests and running at
different speeds, this is painful.

This modifies the test reporters to show the number of tests left. This
gives you, at a glance, which browsers are done and how close each is to
completion.